fix: move ApplyOptionField trait definition to a common place instead of generating everytime
